Move things around a little.
Revert.


Minor.
