+++ libs/utils/Cargo.toml
@@ -15,6 +15,5 @@ proxy-wasm = { version = "0.1.*", features = ["wee_alloc"] }
 indexmap = { version = "1.6.1", features = ["serde-1"] }
 petgraph = {version =  "0.*", features = ["serde-1"] }
 serde = { version = "1.0", features = ["derive"] }
-serde_json = "*"
 pathfinding = "2.1.1"
