+++ rs_filter/filter_base.rs
@@ -9,12 +9,12 @@ use proxy_wasm::types::Action;
 use proxy_wasm::types::ContextType;
 use proxy_wasm::types::LogLevel;
 use std::convert::TryFrom;
+use std::fmt;
 use std::time::Duration;
 
 use utils::graph::iso::find_mapping_shamir_centralized;
 use utils::graph::serde::FerriedData;
 use utils::graph::serde::Property;
-use utils::misc::headers::*;
 
 // These are generated by the filter
 use super::filter::collect_envoy_properties;
@@ -23,6 +23,61 @@ use super::filter::execute_udfs_and_check_trace_lvl_prop;
 use super::filter::get_value_for_storage;
 
 // ---------------------- General Helper Functions ----------------------------
+
+#[repr(i64)]
+#[derive(Debug, PartialEq)]
+pub enum TrafficDirection {
+    Unspecified = 0,
+    Inbound = 1,
+    Outbound = 2,
+}
+impl From<i64> for TrafficDirection {
+    fn from(orig: i64) -> Self {
+        match orig {
+            0x1 => return TrafficDirection::Inbound,
+            0x2 => return TrafficDirection::Outbound,
+            _ => return TrafficDirection::Unspecified,
+        };
+    }
+}
+
+impl fmt::Display for TrafficDirection {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match *self {
+            TrafficDirection::Unspecified => write!(f, "unspecified"),
+            TrafficDirection::Inbound => write!(f, "inbound"),
+            TrafficDirection::Outbound => write!(f, "outbound"),
+        }
+    }
+}
+
+#[repr(u8)]
+#[derive(Debug, PartialEq)]
+#[allow(dead_code)]
+pub enum HttpType {
+    Unspecified = 0,
+    Request = 1,
+    Response = 2,
+}
+
+pub fn data_to_str(stored_data: &FerriedData) -> Option<String> {
+    let stored_data_str: String;
+    match serde_json::to_string(&stored_data) {
+        Ok(stored_data_str_) => {
+            stored_data_str = stored_data_str_;
+        }
+        Err(e) => {
+            log::error!("Could not translate stored data to json string: {:?}\n", e);
+            return None;
+        }
+    }
+    return Some(stored_data_str);
+}
+
+pub fn join_str(str_vec: &Vec<&str>) -> String {
+    return str_vec.join(".");
+}
+
 fn fetch_data_from_headers(ctx: &HttpHeaders, request_type: HttpType) -> FerriedData {
     let data_str_opt: Option<String>;
     if request_type == HttpType::Request {
@@ -40,7 +95,7 @@ fn fetch_data_from_headers(ctx: &HttpHeaders, request_type: HttpType) -> Ferried
                 return fd;
             }
             Err(e) => {
-                log::error!("Could not translate stored data to json string: {0}\n", e);
+                log::error!("Could not translate stored data to json string: {:?}\n", e);
             }
         }
     }
@@ -188,7 +243,7 @@ impl Context for HttpHeaders {
             log::warn!("Storage body: {:?}", body);
         }
         for (name, value) in &self.get_http_response_headers() {
-            log::warn!("Storage Header - {}: {}", name, value);
+            log::warn!("Storage Header - {:?}: {:?}", name, value);
         }
     }
 }
